% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyEmbedding_utils.R
\name{allEqual}
\alias{allEqual}
\title{allEqual}
\usage{
allEqual(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
logical whether a a vector is equal to its first element.
}
\description{
Checks if a vector is equal to its first element
}
\keyword{internal}
