% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyEmbedding_utils.R
\name{buildLabelsDataFrame}
\alias{buildLabelsDataFrame}
\title{buildLabelsDataFrame}
\usage{
buildLabelsDataFrame(labels, resubstituted_labels, k_adaptive)
}
\arguments{
\item{labels}{Is a named character vector with true labels.}

\item{resubstituted_labels}{Is a named character vector with predicted
labels.}

\item{k_adaptive}{Is a named vector of the k-values, this could be a single
integer when fixed.}
}
\value{
A dataframe with rows the same as resubstituted_labels and columns
for input_labels, predicted_labels, and resubstituted_labels.
}
\description{
Build dataframe for output from `classifyEmbedding`
}
\keyword{internal}
