% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyEmbedding_utils.R
\name{combineBinaryErrors}
\alias{combineBinaryErrors}
\title{combineBinaryErrors}
\usage{
combineBinaryErrors(E_list)
}
\arguments{
\item{E_list}{A list containing matrices. Each matrix must have the same
number of columns (k-values) and contain rownames (cells).}
}
\value{
A sparse error matrix.
}
\description{
Combines binary error matrices by averaging error values across all matrices,
for each entry (row and column combination)
}
\keyword{internal}
