% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runOps.R
\name{.runOps}
\alias{.runOps}
\title{runOps}
\usage{
.runOps(obj, ops, leftToRight = TRUE)
}
\arguments{
\item{obj}{obj}

\item{ops}{ops}

\item{leftToRight}{leftToRight}
}
\value{
matrix
}
\description{
runOps
}
\keyword{internal}
