% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureNetwork.R
\name{featureNetwork}
\alias{featureNetwork}
\title{featureNetwork}
\usage{
featureNetwork(assay_list)
}
\arguments{
\item{assay_list}{assay list}
}
\value{
igraph
}
\description{
featureNetwork
}
\examples{

}
