% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyEmbedding_utils.R
\name{getBestColumn}
\alias{getBestColumn}
\title{getBestColumn}
\usage{
getBestColumn(E, balanced_labels = NULL)
}
\arguments{
\item{E}{An error matrix.}

\item{balanced_labels}{Class labels for each row (cell) of E.}
}
\value{
The index of the best performing column of E
}
\description{
Identifies the index of the column of a matrix with the minimum mean.
If balanced_labels is given then calculate the balanced mean
}
\keyword{internal}
