% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyEmbedding_utils.R
\name{getDensityK}
\alias{getDensityK}
\title{getDensityK}
\usage{
getDensityK(coords, k_values = k_values, dist_maxK = 100)
}
\arguments{
\item{coords}{Is a cells (rows) x dimensions matrix for which distances
should be calculated.}

\item{k_values}{Is a numeric character of maximum k-values to use.}

\item{dist_maxK}{Is the maximum distance to consider for estimating the local
density.}
}
\value{
A list.
}
\description{
getDensityK
}
\keyword{internal}
