% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyEmbedding_utils.R
\name{getModeFirst}
\alias{getModeFirst}
\title{getModeFirst}
\usage{
getModeFirst(x, first)
}
\arguments{
\item{x}{A character or a factor.}

\item{first}{An integer.}
}
\value{
A character of the mode of x.
}
\description{
Identify the mode of x up to the first index
}
\keyword{internal}
