% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddingKNN_helper.R
\name{getQueryK}
\alias{getQueryK}
\title{getQueryK}
\usage{
getQueryK(knn, k_local)
}
\arguments{
\item{knn}{Is a k-nearest neighbour matrix, giving the indices of the 
training set that the query is closest to. Rows are the query cells, columns 
are the NNs, should be a large value. Typically output using 
BiocNeighbors::queryKNN(,,k = max(k_local)).}

\item{k_local}{Is an integer vector length of the reference set, giving the 
local k to use. If k_local is given as a single integer, then that value is 
used as k for all observations.}
}
\value{
An integer vector with local k to use for each query cell.
}
\description{
For each cell in the query data, use the 1NN's adaptive k value 
(of the reference data) to identify the local best k value
}
\keyword{internal}
