% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyEmbedding_utils.R
\name{gm_mean}
\alias{gm_mean}
\title{gm_mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector.}

\item{na.rm}{A logical value indicating whether NA values should be stripped
before calculating the geometric mean.}
}
\value{
A numeric.
}
\description{
Calculate the geometric mean
}
\keyword{internal}
