% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapes-pred-grapes.R
\name{\%pred\%}
\alias{\%pred\%}
\title{Binary operator for model predictions on data}
\usage{
data \%pred\% models
}
\arguments{
\item{data}{is a matrix with rows corresponding to features, and columns
corresponding to cells/observations}

\item{models}{is a list of univariate outcome models with the features as
explanatory variables}
}
\value{
a matrix with rows equal to \code{length(models)} and columns
corresponding to cells/observations
}
\description{
This function performs model predictions via the \code{predict} function
for each column of data.
}
\keyword{internal}
