% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapes-projpred-grapes.R
\name{\%projpred\%}
\alias{\%projpred\%}
\title{Project and/or predict data using feature weights or a LDA model object}
\usage{
a \%projpred\% b
}
\arguments{
\item{a}{a matrix with colnames specified}

\item{b}{a matrix with rownames specified, or a lda model object, or a
list containing a matrix and/or a lda model object.}
}
\value{
matrix
}
\description{
This function takes a data matrix a and, depending on the class of b,
projects the data using feature weights, or predicts new values using
linear discriminant analysis (LDA) model object, or both.
}
\keyword{internal}
