% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapes-times-1-grapes.R
\name{\%*1\%}
\alias{\%*1\%}
\title{Sorted matrix multiplication with intercept column}
\usage{
a \%*1\% b
}
\arguments{
\item{a}{a matrix with rownames specified}

\item{b}{a matrix with colnames specified}
}
\value{
matrix
}
\description{
This function first binds a column filled with 1s named \code{intercept} to
\code{a}, then performs rownames and colnames-aware (\code{\%**\%}) matrix
multiplication with \code{b}.
}
\keyword{internal}
