% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapes-times-times-grapes.R
\name{\%**\%}
\alias{\%**\%}
\title{Sorted matrix multiplication}
\usage{
X \%**\% Y
}
\arguments{
\item{X}{a matrix with colnames specified.}

\item{Y}{a matrix with rownames specified. Alternatively, a list assumed to
contain two objects, a matrix with rownames specified, and a vector of
scaling values for subtraction.}
}
\value{
matrix
}
\description{
This function multiplies two matrices but first reorders the rows of the
second matrix to match the columns of the first matrix
}
\keyword{internal}
