% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeEmbedding.R
\name{imputeEmbedding}
\alias{imputeEmbedding}
\title{imputeEmbedding}
\usage{
imputeEmbedding(
  assay_list,
  embedding = NULL,
  reference = Reduce(assay_list, colnames),
  query = Reduce(assay_list, colnames),
  neighbours = 5,
  fun = mean,
  ...
)
}
\arguments{
\item{assay_list}{assay_list}

\item{embedding}{embedding}

\item{reference}{reference}

\item{query}{query}

\item{neighbours}{neighbours}

\item{fun}{fun}

\item{...}{...}
}
\value{
matrix
}
\description{
imputeEmbedding
}
\examples{

}
