% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyEmbedding_utils.R
\name{isUnequal}
\alias{isUnequal}
\title{isUnequal}
\usage{
isUnequal(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
An integer vector. 1 for unequal. 0 for equal
}
\description{
Checks if elements of 2 vectors are unequal
}
\keyword{internal}
