% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaicDataUpSet.R
\name{mosaicDataUpSet}
\alias{mosaicDataUpSet}
\title{mosaicDataUpSet}
\usage{
mosaicDataUpSet(assay_list, plot = TRUE)
}
\arguments{
\item{assay_list}{a list of matrices with rownames (features) specified}
}
\value{
UpSet object displaying degree of overlap of rownames (features)
among each of the data matrices in \code{assay_list}.
}
\description{
plots feature overlaps as an UpSet plot
}
\examples{

}
