% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryNamedKNN.R
\name{queryNamedKNN}
\alias{queryNamedKNN}
\title{queryNamedKNN}
\usage{
queryNamedKNN(coords_reference, coords_query, k)
}
\arguments{
\item{coords_reference}{coords_reference}

\item{coords_query}{coords_query}

\item{k}{k}
}
\value{
matrix
}
\description{
queryNamedKNN
}
\keyword{internal}
