% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reWeightEmbedding.R
\name{reWeightEmbedding}
\alias{reWeightEmbedding}
\title{Re-weight embedding}
\usage{
reWeightEmbedding(embedding, weights = NULL, factor = 1e+06)
}
\arguments{
\item{embedding}{Output from stabMap}

\item{weights}{weights vector}

\item{factor}{numeric value}
}
\value{
matrix
}
\description{
Re-weights embedding according to given weights
}
\examples{

}
