% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabMap_utils.R
\name{selectFeatures}
\alias{selectFeatures}
\title{selectFeatures}
\usage{
selectFeatures(assay, features, maxFeatures)
}
\arguments{
\item{assay}{An assay matrix rows are features, columns are cells}

\item{features}{Character vector of the current features that are selected}

\item{maxFeatures}{Integer of the number of maxFeatures to select}
}
\value{
A character vector of the selected features according to variance
ranking.
}
\description{
For a given assay and set of features, perform variance ranking and select
a subset of features
}
\keyword{internal}
