% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyEmbedding_utils.R
\name{smoothLocal}
\alias{smoothLocal}
\title{smoothLocal}
\usage{
smoothLocal(best_k, local, smooth = 10, mean_type = "geometric")
}
\arguments{
\item{best_k}{Is a named vector of local best k values}

\item{local}{Is a KNN matrix, with rows same as best_k and values indices of
best_k.}

\item{smooth}{An integer of k-nearest neighbours to smooth over.}

\item{mean_type}{Character indicating to calculate the 'geometric' or
'arithmetic' mean.}
}
\value{
A numeric vector of smoothed adaptive k-values.
}
\description{
Smooth out the adaptive k values. Can be smoothed by computing the
arithmetic or geometric mean of the adaptive k-values for each cells
neighbourhood
}
\keyword{internal}
