% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseCoef.R
\name{sparseCoef}
\alias{sparseCoef}
\title{sparseCoef}
\usage{
sparseCoef(PCs_X, PCs_X_genes)
}
\arguments{
\item{PCs_X}{PCs_X}

\item{PCs_X_genes}{PCs_X_genes}
}
\value{
matrix
}
\description{
sparseCoef
}
\examples{

}
