% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabMap.R
\name{stabMap}
\alias{stabMap}
\title{stabMap}
\usage{
stabMap(
  assay_list,
  labels_list = NULL,
  reference_list = sapply(names(assay_list), function(x) TRUE, simplify = FALSE),
  reference_features_list = lapply(assay_list, rownames),
  ncomponentsReference = 50,
  ncomponentsSubset = 50,
  suppressMessages = TRUE,
  projectAll = FALSE,
  maxFeatures = 1000,
  plot = TRUE,
  scale.center = FALSE,
  scale.scale = FALSE
)
}
\arguments{
\item{assay_list}{assay_list}

\item{labels_list}{labels_list}

\item{reference_list}{reference_list}

\item{reference_features_list}{reference_features_list}

\item{ncomponentsReference}{ncomponentsReference}

\item{ncomponentsSubset}{ncomponentsSubset}

\item{suppressMessages}{suppressMessages}

\item{projectAll}{projectAll}

\item{maxFeatures}{maxFeatures}

\item{plot}{plot}

\item{scale.center}{scale.center}

\item{scale.scale}{scale.scale}
}
\value{
matrix
}
\description{
stabMap
}
\examples{

# simulate some data, 300 genes, 150 cells.
full_expr = matrix(rnorm(300*150),
nrow = 300, ncol = 150,
dimnames = list(paste0("gene_", 1:300),
paste0("cell_", 1:150)))

# build list of discrete assays with non-overlapping features
assay_list = list(
D1 = full_expr[1:150, 1:50],
D2 = full_expr[76:215, 51:100],
D3 = full_expr[151:300, 101:150]
)

# assign labels to one group of cells
labels_list = list(
D1 = rep(letters[1:5], each = 10)
)

# note which datasets should be treated as references
reference_list = list(
D1 = TRUE,
D2 = FALSE,
D3 = TRUE
)

# examine the topology of this mosaic data integration
mosaicDataUpSet(assay_list)
plot(mosaicDataTopology(assay_list))

# perform stabMap
out = stabMap(assay_list,
labels_list = labels_list,
reference_list = reference_list,
ncomponentsReference = 20,
ncomponentsSubset = 20)

head(out)

}
