% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorSubset.R
\name{vectorSubset}
\alias{vectorSubset}
\title{vectorSubset}
\usage{
vectorSubset(vec, mat)
}
\arguments{
\item{vec}{vec}

\item{mat}{mat}
}
\value{
matrix
}
\description{
vectorSubset
}
\keyword{internal}
