% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runOps.R
\name{.runOps}
\alias{.runOps}
\title{Run a sequence of binary operations}
\usage{
.runOps(obj, ops, leftToRight = TRUE)
}
\arguments{
\item{obj}{list of objects.}

\item{ops}{list of operations (length should be 1 less than `obj`).}

\item{leftToRight}{logical whether operations should be performed in order
from left to right (default), or right to left.}
}
\value{
matrix or array output of the sequence of binary operations
}
\description{
Run a sequence of binary operations
}
\keyword{internal}
