% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaicDataTopology.R
\name{mosaicDataTopology}
\alias{mosaicDataTopology}
\title{mosaicDataTopology}
\usage{
mosaicDataTopology(assay_list)
}
\arguments{
\item{assay_list}{a list of data matrices with rownames (features) specified.}
}
\value{
igraph weighted network with nodes corresponding to
\code{assay_list} elements, and edges present if the matrices share at
least one rowname. Edge weights correspond to the number of shared
rownames among data matrices.
}
\description{
Generate mosaic data topology network as an igraph object.
}
\examples{
set.seed(2021)
assay_list = mockMosaicData()
mdt = mosaicDataTopology(assay_list)
mdt
plot(mdt)

}
