% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabMap.R
\name{stabMap}
\alias{stabMap}
\title{stabMap}
\usage{
stabMap(
  assay_list,
  labels_list = NULL,
  reference_list = sapply(names(assay_list), function(x) TRUE, simplify = FALSE),
  reference_features_list = lapply(assay_list, rownames),
  ncomponentsReference = 50,
  ncomponentsSubset = 50,
  suppressMessages = TRUE,
  projectAll = FALSE,
  maxFeatures = 1000,
  plot = TRUE,
  scale.center = FALSE,
  scale.scale = FALSE
)
}
\arguments{
\item{assay_list}{assay_list}

\item{labels_list}{labels_list}

\item{reference_list}{reference_list}

\item{reference_features_list}{reference_features_list}

\item{ncomponentsReference}{ncomponentsReference}

\item{ncomponentsSubset}{ncomponentsSubset}

\item{suppressMessages}{suppressMessages}

\item{projectAll}{projectAll}

\item{maxFeatures}{maxFeatures}

\item{plot}{plot}

\item{scale.center}{scale.center}

\item{scale.scale}{scale.scale}
}
\value{
matrix
}
\description{
stabMap
}
\examples{

}
